--- a/attrs_bsd.go
+++ b/attrs_bsd.go
@@ -1,5 +1,4 @@
 // +build darwin freebsd netbsd
-// +build cgo
 
 package sftp