--- a/attrs_stubs.go
+++ b/attrs_stubs.go
@@ -1,4 +1,4 @@
-// +build !cgo plan9 windows android
+// +build plan9 windows android
 
 package sftp