--- a/attrs_unix.go
+++ b/attrs_unix.go
@@ -1,5 +1,4 @@
 // +build aix dragonfly !android,linux openbsd solaris
-// +build cgo
 
 package sftp