… attrs_test.go
… client_integration_test.go
… client_test.go
… conn_test.go
… internal/encoding/ssh/filexfer/attrs.go
… internal/encoding/ssh/filexfer/attrs_test.go
… internal/encoding/ssh/filexfer/escape_analysis_test.go
… internal/encoding/ssh/filexfer/extended_packets_test.go
… internal/encoding/ssh/filexfer/extensions.go
… internal/encoding/ssh/filexfer/extensions_test.go
… internal/encoding/ssh/filexfer/filexfer.go
… internal/encoding/ssh/filexfer/fx.go
… internal/encoding/ssh/filexfer/fx_test.go
… internal/encoding/ssh/filexfer/handle_packets.go
… internal/encoding/ssh/filexfer/handle_packets_test.go
… internal/encoding/ssh/filexfer/init_packets.go
… internal/encoding/ssh/filexfer/init_packets_test.go
… internal/encoding/ssh/filexfer/open_packets.go
… internal/encoding/ssh/filexfer/open_packets_test.go
… internal/encoding/ssh/filexfer/openssh/fsync.go
… internal/encoding/ssh/filexfer/openssh/fsync_test.go
… internal/encoding/ssh/filexfer/openssh/hardlink.go
… internal/encoding/ssh/filexfer/openssh/hardlink_test.go
… internal/encoding/ssh/filexfer/openssh/posix-rename.go
… internal/encoding/ssh/filexfer/openssh/posix-rename_test.go
… internal/encoding/ssh/filexfer/openssh/statvfs.go
… internal/encoding/ssh/filexfer/openssh/statvfs_test.go
… internal/encoding/ssh/filexfer/packets.go
… internal/encoding/ssh/filexfer/packets_test.go
… internal/encoding/ssh/filexfer/path_packets.go