So, I’m checking in and pushing a PR very early here, so that others can get a sense of how the client refactor will look like after integrating the filexfer.

The early attention here should be on the `Client.close()` function, which felt would be the quickest way to show off how things would change. The changes to `Conn` and `clientConn` are to support the changes in the `close()` function, and allow for a parallel old-way/new-way hand-off from one set of wheels to the other while the car is still running. Before the PR is merged, it will eventually only have the new-way.