@puellanivis 2021-04-24
From:
```
sftp.clean$ go tool pprof memprofile.out 
(pprof) top10
Showing nodes accounting for 500.49GB, 98.94% of 505.86GB total
Dropped 170 nodes (cum <= 2.53GB)
Showing top 10 nodes out of 28
      flat  flat%   sum%        cum   cum%
  403.43GB 79.75% 79.75%   403.43GB 79.75%  github.com/pkg/sftp.recvPacket
   40.09GB  7.92% 87.68%    40.09GB  7.92%  github.com/pkg/sftp.(*bufPool).Get
   15.94GB  3.15% 90.83%    15.94GB  3.15%  github.com/pkg/sftp.(*sshFxpWritePacket).marshalPacket
   15.48GB  3.06% 93.89%    15.48GB  3.06%  github.com/pkg/sftp.(*allocator).GetPage
    8.21GB  1.62% 95.51%     8.21GB  1.62%  github.com/pkg/sftp.(*sshFxInitPacket).MarshalBinary
    7.27GB  1.44% 96.95%     7.27GB  1.44%  github.com/pkg/sftp.(*sshFxpOpenPacket).MarshalBinary
    5.20GB  1.03% 97.97%     5.25GB  1.04%  github.com/pkg/sftp.benchmarkWriteTo
    2.19GB  0.43% 98.41%     3.55GB   0.7%  github.com/pkg/sftp.benchmarkReadFrom
    2.10GB  0.42% 98.82%     3.59GB  0.71%  github.com/pkg/sftp.benchmarkWrite
    0.59GB  0.12% 98.94%     2.69GB  0.53%  github.com/pkg/sftp.(*File).writeChunkAt
```
to:
```
sftp$ go tool pprof memprofile.out 
(pprof) top10
Showing nodes accounting for 93.60GB, 98.37% of 95.16GB total
Dropped 130 nodes (cum <= 0.48GB)
Showing top 10 nodes out of 42
      flat  flat%   sum%        cum   cum%
   54.95GB 57.75% 57.75%    54.95GB 57.75%  github.com/pkg/sftp.(*bufPool).Get
   16.50GB 17.34% 75.08%    16.50GB 17.34%  github.com/pkg/sftp.(*allocator).GetPage
    8.24GB  8.66% 83.74%     8.24GB  8.66%  github.com/pkg/sftp.(*sshFxInitPacket).MarshalBinary
    5.10GB  5.36% 89.11%     5.15GB  5.41%  github.com/pkg/sftp.benchmarkWriteTo
    2.21GB  2.32% 91.43%    28.50GB 29.95%  github.com/pkg/sftp.(*clientConn).sendPacket
    2.19GB  2.30% 93.73%     3.61GB  3.79%  github.com/pkg/sftp.benchmarkReadFrom
    1.97GB  2.07% 95.80%     3.05GB  3.21%  github.com/pkg/sftp.benchmarkWrite
    0.86GB   0.9% 96.71%     0.86GB   0.9%  github.com/pkg/sftp.(*delayedWriter).Write
    0.80GB  0.84% 97.54%    27.30GB 28.69%  github.com/pkg/sftp.(*File).readChunkAt
    0.78GB  0.82% 98.37%     0.78GB  0.82%  github.com/pkg/sftp.(*File).WriteTo.func2
```